# gravitee-rest-api-portal/gravitee-rest-api-portal-rest/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Gravitee.io Portal Rest API
  description: API dedicated to the devportal part of Gravitee
  termsOfService: terms
  contact:
    name: The Gravitee Team
    url: https://gravitee.io
    email: contact@graviteesource.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: "3.0.0"
servers:
  - url: http://localhost:8083/portal/environments/{envId}
    description: The portal API for a given environment
    variables:
      envId:
        default: DEFAULT
        description: ID of the environment
...
security:
  - BasicAuth: []
  - CookieAuth: []
paths:
...
  /subscriptions:
    get:
      tags:
        - Subscription
      parameters:
        - $ref: '#/components/parameters/apiIdQueryParam'
        - $ref: '#/components/parameters/applicationIdQueryParam'
        - $ref: '#/components/parameters/subscriptionStatusesQueryParam'
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
      description: |
        List all ACCEPTED, PAUSED & PENDING subscriptions, filtered by api and/or by application. At least an api or an application must be provided.

        User must have the APPLICATION_SUBSCRIPTION[READ] permission to list subscription with application query param.\
        User must have the API_SUBSCRIPTION[READ] permission to list subscription with api query param.

      operationId: getSubscriptions
      responses:
        200:
          description: List of Subscriptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionsResponse'
        400:
          description: At least an api or an application must be provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          $ref: '#/components/responses/PermissionError'
        500:
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Subscription
      requestBody:
        description: Use to create a subscription.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionInput'
      summary: Create a subscription.
      description: |
        Create a new subscription.

        User must have APPLICATION_SUBSCRIPTION[CREATE] permission.
      operationId: createSubscription
      responses:
        200:
          description: Created subscription.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        400:
          description: Input must not be null.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          $ref: '#/components/responses/PermissionError'
        500:
          $ref: '#/components/responses/InternalServerError'
  /subscriptions/{subscriptionId}:
    parameters:
      - $ref: '#/components/parameters/subscriptionIdParam'
    get:
      tags:
        - Subscription
      parameters:
        - $ref: '#/components/parameters/subscriptionIncludeQueryParam'

      summary: Get a subscription.
      description: |
        Get a subscription.

        User must have API_SUBSCRIPTION[CREATE] or APPLICATION_SUBSCRIPTION[CREATE] permission.
      operationId: getSubscriptionById
      responses:
        200:
          description: One subscription.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  parameters:
...
    subscriptionIdParam:
      name: subscriptionId
      in: path
      required: true
      description: Id of a subscription.
      schema:
        type: string
...
    apiIdQueryParam:
      name: apiId
      in: query
      required: false
      description: Id of an api.
      schema:
        type: string
    applicationIdQueryParam:
      name: applicationId
      in: query
      required: false
      description: Id of an application.
      schema:
        type: string
...
    pageNumberParam:
      name: page
      in: query
      required: false
      description: The page number for pagination.
      schema:
        type: integer
        default: 1
    pageSizeParam:
      name: size
      in: query
      required: false
      description: |
        The number of items per page for pagination.
        If the size is 0, the response contains only metadata and returns the values as for a non-paged resource.
        If the size is -1, the response contains all datas.
      schema:
        type: integer
        default: 10

    #Apis filtering
...
    subscriptionStatusesQueryParam:
      name: statuses
      in: query
      required: false
      description: status of subscription.
      schema:
        type: array
        items:
          type: string
          enum:
            - ACCEPTED
            - CLOSED
            - PAUSED
            - PENDING
            - REJECTED
    subscriptionIncludeQueryParam:
      name: include
      in: query
      required: false
      description: Comma-separated list of related objects to include in the response.
      schema:
        type: array
        items:
          type: string
          enum:
            - keys

    #Users filtering
...
  schemas:
...
    ErrorResponse:
      properties:
        errors:
          description: List of errors
          type: array
          items:
            $ref: '#/components/schemas/Error'
...
    SubscriptionsResponse:
      properties:
        data:
          description: List of subscriptions.
          type: array
          items:
            $ref: '#/components/schemas/Subscription'
        metadata:
          $ref: '#/components/schemas/MetadataMap'
        links:
          $ref: '#/components/schemas/Links'
...
    Error:
      properties:
        status:
          description: Status HTTP of the error
          type: string
        message:
          description: Message of the error
          type: string
        code:
          description: Technical code of the error
          type: string
        parameters:
          description: Parameters
          type: object
          additionalProperties:
            type: string
...
    Subscription:
      required:
        - id
        - api
        - application
        - plan
        - status
      properties:
        id:
          description: Unique identifier of a subscription.
          type: string
        api:
          description: Subscribed API.
          type: string
        application:
          description: Subscribing application.
          type: string
        plan:
          description: Subscribed plan.
          type: string
        request:
          description: Comment of the user when asking for a subscription.
          type: string
        reason:
          description: Reason of the user when processing (accepting/rejecting) a subscription.
          type: string
        created_at:
          description: Creation date and time of the subscription.
          type: string
          format: date-time
        processed_at:
          description: Date and time when the subscription request was processed.
          type: string
          format: date-time
        start_at:
          description: Start date and time of the subscription.
          type: string
          format: date-time
        end_at:
          description: Expiration date and time of the subscription.
          type: string
          format: date-time
        paused_at:
          description: Paused date and time of the subscription.
          type: string
          format: date-time
        closed_at:
          description: Closed date and time of the subscription.
          type: string
          format: date-time
        subscribed_by:
          description: The user who subscribed.
          type: string
        status:
          description: Status of the subscription.
          type: string
          enum:
            - PENDING
            - ACCEPTED
            - CLOSED
            - REJECTED
            - PAUSED
        ###################
        # may be included #
        ###################
        keys:
          description: |
            Only returned with */subscriptions/{subscriptionId}*.
            Need *include* query param to contain 'keys'.

            List of APIKeys of the subscription.
          type: array
          items:
            $ref: '#/components/schemas/Key'
...
    SubscriptionInput:
      properties:
        application:
          description: Id of the application which subscribe
          type: string
        plan:
          description: Id of the plan to which the application subscribe
          type: string
        request:
          description: A request message to the api owner why a user want to subscribe
          type: string
...
    Key:
      description: Subscription Key
      properties:
        id:
          type: string
        api:
          type: string
        application:
          type: string
        plan:
          type: string
        paused:
          type: boolean
        revoked:
          type: boolean
        expired:
          type: boolean
        created_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
        expire_at:
          type: string
          format: date-time
...
    MetadataMap:
      description: Map of Map of Object
      type: object
      additionalProperties:
        type: object
        additionalProperties:
          type: object
...
    Links:
      properties:
        self:
          type: string
        first:
          type: string
        last:
          type: string
        prev:
          type: string
        next:
          type: string
...
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PermissionError:
      description: Insufficient rights to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    CookieAuth:
      type: apiKey
      in: cookie
      name: Auth-Graviteeio-APIM